Add files via upload
main program that asks the user for an array size and a maximum value. The program
should then generate the array using randomarray() , print it to the screen, then sort it using
bubblesort() and print it again. The Two runs with the exact same inputs should produce the same
results.